#811: lazy-load handlebars/nunjucks
