+++ src/blocks/combinators.ts
@@ -16,7 +16,7 @@
  */
 
 import blockRegistry from "@/blocks/registry";
-import { engineRenderer, mapArgs } from "@/helpers";
+import { mapArgs } from "@/helpers";
 import ArrayCompositeReader from "@/blocks/readers/ArrayCompositeReader";
 import CompositeReader from "@/blocks/readers/CompositeReader";
 import { locate } from "@/background/locator";
@@ -51,6 +51,7 @@ import {
   InputValidationError,
   PipelineConfigurationError,
 } from "@/blocks/errors";
+import { engineRenderer } from "@/utils/renderers";
 
 export type ReaderConfig =
   | string
@@ -191,8 +192,10 @@ async function runStage(
     }
   } else {
     // HACK: hack to avoid applying a list to the config for blocks that pass a list to the next block
+    const renderer = await engineRenderer(stage.templateEngine);
+
     blockArgs = isPlainObject(args)
-      ? mapArgs(stageConfig, argContext, engineRenderer(stage.templateEngine))
+      ? mapArgs(stageConfig, argContext, renderer)
       : stageConfig;
 
     if (logValues) {
@@ -326,10 +329,12 @@ export async function reducePipeline(
       const stageRoot = $stageRoot.get(0);
 
       if ("if" in stage) {
+        const renderer = await engineRenderer(stage.templateEngine);
+
         const { if: condition } = mapArgs(
           { if: stage.if },
           { ...extraContext, ...currentArgs },
-          engineRenderer(stage.templateEngine)
+          renderer
         );
         if (!boolean(condition)) {
           logger.debug(
